requirements.txt
google-generativeai>=0.8.0
pydantic>=2.0.0
httpx>=0.27.0
python-slugify>=8.0.0
jinja2>=3.1.0
markdown>=3.6
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dateutil>=2.9.0
rich>=13.0.0
